app/requirements.txt
fastapi
uvicorn
psycopg2-binary
jinja2
python-dotenv
python-multipart
bcrypt
pyvis
iputils
itsdangerous